Puf - add docs for (enhanced) Firebase storage library (#346)
* Create storage-library.md

First commit of documentation for the Firebase Storage library.

* Extract inline image into file

Extract inline image into file in the right location and then link it there.

I included both the original png and a derived svg in there, but am linking only the svg from the markdown as that one renders more readably.

* Update storage-library.md

Remove image as one was too vague and the other too different. The same info is in the descriptions of the custom actions also.

* update formatting

* Remove unused files

---------

Co-authored-by: PoojaB26 <pbhaumik26@gmail.com>